refactor: SVG層の f32 を f64 に統一し DB層との型不整合を解消
- SvgOptions の opacity, blur_radius を f64 に変更
- svg_builder 内の定数・ローカル変数を f64 に統一
- renderer/mod.rs の as f32 キャストを削除

+++ src/renderer/mod.rs
@@ -175,9 +175,9 @@ impl Renderer {
             bg_color: &bg_color,
             language,
             title_bar_style: &render_options.title_bar_style,
-            opacity: render_options.opacity as f32,
+            opacity: render_options.opacity,
             background_image: bg_image_base64.as_deref(),
-            blur_radius: render_options.blur_radius as f32,
+            blur_radius: render_options.blur_radius,
             max_line_length: render_options.max_line_length,
             show_line_numbers: render_options.show_line_numbers,
         };
